[styleguide] - move welcome to own page to clean up routes file
