+++ frontend/src/metabase/internal/pages/WelcomePage.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+const WelcomePage = () => {
+  return (
+    <div className="wrapper">
+      <div className="my4">
+        <h1>Metabase Style Guide</h1>
+        <p className="text-paragraph">
+          Reference and samples for how to make things the Metabase way.
+        </p>
+      </div>
+    </div>
+  );
+};
+
+export default WelcomePage;
